refactor(API): Extracts postJwtLoginRequest to request handlers.
